Fix stop ID overflow, null safety, stdin, and departures response shape
- Change stop/departure IDs from Int to Long; Transitland IDs exceed
  Int.MAX_VALUE (e.g. 2167666560)
- Make StopJson fields nullable to guard against Gson bypassing Kotlin
  null safety; filter incomplete stops via mapNotNull
- Wire Gradle run task stdin so the CLI blocks for input
- Fix DeparturesResponse: API wraps results in stops[], not a top-level
  departures[]; parent stations carry departures on child platform stops,
  so flatten via recursive allDepartures()
- Add JUnit 5 integration tests for getNearbyStops and getDepartures
  using hardcoded SF Caltrain coordinates

+++ src/main/kotlin/dev/catchthenext/model/Departure.kt
@@ -1,7 +1,7 @@
 package dev.catchthenext.model
 
 data class Departure(
-    val stopId: Int,
+    val stopId: Long,
     val departureTime: String,
     val departureMinutes: Long,
     val routeShortName: String,
